backend/trip-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
requests
